+++ stroom-query-common/src/main/java/stroom/query/common/v2/TableResultCreator.java
@@ -17,6 +17,7 @@
 package stroom.query.common.v2;
 
 import stroom.dashboard.expression.v1.Generator;
+import stroom.dashboard.expression.v1.Selector;
 import stroom.dashboard.expression.v1.Val;
 import stroom.query.api.v2.Field;
 import stroom.query.api.v2.OffsetRange;
@@ -112,6 +113,8 @@ public class TableResultCreator implements ResultCreator {
         final Items<Item> items = data.getChildMap().get(parentKey);
         if (items != null) {
             for (final Item item : items) {
+                final GroupKey groupKey = item.getKey();
+
                 // If the result is within the requested window (offset + length) then add it.
                 if (pos >= offset &&
                         resultList.size() < length) {
@@ -125,8 +128,34 @@ public class TableResultCreator implements ResultCreator {
                         if (item.getGenerators().length > i) {
                             final Generator generator = item.getGenerators()[i];
                             if (generator != null) {
-                                // Convert all list into fully resolved objects evaluating functions where necessary.
-                                final Val val = generator.eval();
+                                Val val;
+
+                                if (groupKey != null && generator instanceof Selector) {
+                                    // If the generator is a selector then select a child row.
+                                    final Items<Item> childItems = data.getChildMap().get(groupKey);
+                                    if (childItems != null) {
+                                        // Create a list of child generators.
+                                        final List<Generator> childGenerators = new ArrayList<>(childItems.size());
+                                        for (final Item childItem : childItems) {
+                                            final Generator childGenerator = childItem.getGenerators()[i];
+                                            childGenerators.add(childGenerator);
+                                        }
+
+                                        // Make the selector select from the list of child generators.
+                                        final Selector selector = (Selector) generator;
+                                        val = selector.select(childGenerators.toArray(new Generator[0]));
+
+                                    } else {
+                                        // If there are are no child items then just evaluate the inner expression
+                                        // provided to the selector function.
+                                        val = generator.eval();
+                                    }
+                                } else {
+                                    // Convert all list into fully resolved objects evaluating functions where
+                                    // necessary.
+                                    val = generator.eval();
+                                }
+
                                 string = fieldFormatter.format(field, val);
                             }
                         }
@@ -146,7 +175,7 @@ public class TableResultCreator implements ResultCreator {
                 pos++;
 
                 // Add child results if a node is open.
-                if (item.getKey() != null && openGroups != null && openGroups.contains(item.getKey().toString())) {
+                if (groupKey != null && openGroups != null && openGroups.contains(item.getKey().toString())) {
                     pos = addTableResults(data, fields, maxResults, offset, length, openGroups, resultList,
                             item.getKey(), depth + 1, pos);
                 }